requirements.txt
azure-functions==1.7.0
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
azure-cosmos==4.0.0
requests==2.25.1
python-dotenv==0.17.1
pymongo==4.8.0